fix ub in BloomFilter

+++ src/paimon/common/utils/bloom_filter.cpp
@@ -61,7 +61,9 @@ Status BloomFilter::AddHash(int32_t hash1) {
     auto hash2 = static_cast<int32_t>(static_cast<uint32_t>(hash1) >> 16);
 
     for (int32_t i = 1; i <= num_hash_functions_; i++) {
-        int32_t combined_hash = hash1 + (i * hash2);
+        // Use uint32_t arithmetic to avoid signed overflow UB (matches Java int wrap semantics)
+        int32_t combined_hash = static_cast<int32_t>(
+            static_cast<uint32_t>(hash1) + (static_cast<uint32_t>(i) * static_cast<uint32_t>(hash2)));
         // hashcode should be positive, flip all the bits if it's negative
         if (combined_hash < 0) {
             combined_hash = ~combined_hash;
@@ -76,7 +78,9 @@ bool BloomFilter::TestHash(int32_t hash1) const {
     auto hash2 = static_cast<int32_t>(static_cast<uint32_t>(hash1) >> 16);
 
     for (int32_t i = 1; i <= num_hash_functions_; i++) {
-        int32_t combined_hash = hash1 + (i * hash2);
+        // Use uint32_t arithmetic to avoid signed overflow UB (matches Java int wrap semantics)
+        int32_t combined_hash = static_cast<int32_t>(
+            static_cast<uint32_t>(hash1) + (static_cast<uint32_t>(i) * static_cast<uint32_t>(hash2)));
         // hashcode should be positive, flip all the bits if it's negative
         if (combined_hash < 0) {
             combined_hash = ~combined_hash;
